end tests

+++ Common/Snapshot/StatsDto.cpp
@@ -1,7 +1,7 @@
 #include "StatsDto.h"
 
 StatsDto::StatsDto(const int8_t& playerId, const int16_t& kills, const int32_t& rankingKills,
-    const int16_t shots, const int32_t& rankingShots, const int8_t& minutes, const int8_t& seconds,
+    const int16_t shots, const int32_t& rankingShots, const int16_t& minutes, const int16_t& seconds,
     const int32_t& rankingDuration) : playerId(playerId), kills(kills), rankingKills(rankingKills),
     shots(shots), rankingShots(rankingShots), minutes(minutes), seconds(seconds), rankingDuration(rankingDuration) {}
 
@@ -20,11 +20,11 @@ int16_t StatsDto::getShots() const {
     return shots;
 }
 
-int8_t StatsDto::getMinutes() const {
+int16_t StatsDto::getMinutes() const {
     return minutes;
 }
 
-int8_t StatsDto::getSeconds() const {
+int16_t StatsDto::getSeconds() const {
     return seconds;
 }
 
@@ -40,11 +40,11 @@ int32_t StatsDto::getRankingDuration() const {
     return rankingDuration;
 }
 
-void StatsDto::setMinutes(const int8_t& inputMinutes) {
+void StatsDto::setMinutes(const int16_t& inputMinutes) {
     minutes = inputMinutes;
 }
 
-void StatsDto::setSeconds(const int8_t& inputSeconds) {
+void StatsDto::setSeconds(const int16_t& inputSeconds) {
     seconds = inputSeconds;
 }
