+++ Server/src/GameThread/StatsController.h
@@ -11,7 +11,7 @@ class StatsController {
     private:
     std::vector<int16_t> killsRanking;
     std::vector<int16_t> shotsRanking;
-    std::vector<std::pair<int8_t, int8_t>> durationRanking;
+    std::vector<std::pair<int16_t, int16_t>> durationRanking;
     std::string path;
     std::mutex mutex;
 
@@ -19,13 +19,13 @@ class StatsController {
 
     void insertShotsRanking(const int16_t& shots);
 
-    void insertDurationRanking(const std::pair<int8_t, int8_t>& duration);
+    void insertDurationRanking(const std::pair<int16_t, int16_t>& duration);
 
     uint32_t getKillsIndex(const int16_t& kills);
 
     uint32_t getShotsIndex(const int16_t& shots);
 
-    uint32_t getDurationIndex(const std::pair<int8_t, int8_t>& duration);
+    uint32_t getDurationIndex(const std::pair<int16_t, int16_t>& duration);
 
     public:
     /*
@@ -33,7 +33,7 @@ class StatsController {
      */
     StatsController(const std::string& path);
 
-    std::shared_ptr<Snapshot> updateStats(std::vector<StatsDto> statsFromGame, const int8_t& minutes, const int8_t& seconds);
+    std::shared_ptr<Snapshot> updateStats(std::vector<StatsDto> statsFromGame, const int16_t& minutes, const int16_t& seconds);
 
     ~StatsController();
 };
